Coordinates in Java world are now always GPS-style projection and converted to map internal one Fixed problems in MapClick Added distance calculation to be used to determine the height of the map for example Integrated zoom in datamodel and event payload as well as logic (in JS and Java)

+++ xml/src/main/java/org/geojsf/model/xml/geojsf/Scale.java
@@ -43,7 +43,7 @@ public class Scale
     @XmlAttribute(name = "unit")
     protected String unit;
     @XmlAttribute(name = "zoom")
-    protected Integer zoom;
+    protected Double zoom;
 
     /**
      * Gets the value of the value property.
@@ -101,7 +101,7 @@ public class Scale
      *     {@link Integer }
      *     
      */
-    public int getZoom() {
+    public double getZoom() {
         return zoom;
     }
 
@@ -113,7 +113,7 @@ public class Scale
      *     {@link Integer }
      *     
      */
-    public void setZoom(int value) {
+    public void setZoom(double value) {
         this.zoom = value;
     }
